Refactor to use ApiSignedTransaction in DryrunRequest
