+++ algonaut_model/src/lib.rs
@@ -1,6 +1,4 @@
-/// Algorand protocol daemon
 pub mod algod;
-/// Algorand's indexer
 pub mod indexer;
-/// Key management daemon
 pub mod kmd;
+pub mod transaction;
