+++ examples/logic_sig_delegated.rs
@@ -1,8 +1,9 @@
 use algonaut::algod::v2::Algod;
-use algonaut::core::{LogicSignature, MicroAlgos, SignedLogic};
+use algonaut::core::{LogicSignature, MicroAlgos};
 use algonaut::transaction::transaction::TransactionSignature;
 use algonaut::transaction::{account::Account, TxnBuilder};
 use algonaut::transaction::{Pay, SignedTransaction};
+use algonaut_transaction::transaction::SignedLogic;
 use dotenv::dotenv;
 use std::env;
 use std::error::Error;
