+++ examples/logic_sig_delegated_multi.rs
@@ -1,8 +1,9 @@
 use algonaut::algod::v2::Algod;
-use algonaut::core::{LogicSignature, MicroAlgos, MultisigAddress, SignedLogic};
+use algonaut::core::{LogicSignature, MicroAlgos, MultisigAddress};
 use algonaut::transaction::transaction::TransactionSignature;
 use algonaut::transaction::{account::Account, TxnBuilder};
 use algonaut::transaction::{Pay, SignedTransaction};
+use algonaut_transaction::transaction::SignedLogic;
 use dotenv::dotenv;
 use std::env;
 use std::error::Error;
